# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: ragota-core API
  description: >
    RAG code indexer: repositories, hybrid search, code navigation,
    cross-service graph and LLM retrieval context.

    `version` below is the wire contract, not the build. GET /health reports
    both: `api_version` is this number, `version` is the binary that answered.
  version: 0.2.0
servers:
  - url: /
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /webhooks/git:
    post:
      tags: [repos]
      summary: Git push webhook (GitHub/GitLab)
      description: >
        Finds the repository matching the payload and triggers update + reindex.
        Repositories are matched on the normalized clone URL, then on the exact
        name, then on the final path segment — never on a substring, so a push
        to "gateway" cannot reindex "gateway-v2".

        Outside API-key auth (CI systems cannot set one), so RAGOTA_WEBHOOK_SECRET
        (X-Webhook-Token header or ?token=) is mandatory: with the secret unset
        the endpoint is disabled and answers 503. The route is still rate
        limited and body capped.
      security: []
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitWebhookPayload"
      responses:
        "202":
          description: Reindex triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: indexing
                  repo_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Missing or invalid webhook token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          description: Webhooks are disabled (RAGOTA_WEBHOOK_SECRET is not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/repos:
    get:
      tags: [repos]
      summary: List repositories
      description: >
        Every registered repository with its indexing state. Start here to learn
        the repository ids that `repos` and `repo` parameters elsewhere take,
        and to check `status` and `indexed_at` before reading anything into an
        empty retrieval answer. /stats says whether the indexes hold anything at
        all; /repos/{id}/coverage says whether what they hold is complete.


        `active` is the other field to read before believing an empty answer.
        /search and /context with no `repos` selector answer from the active
        repositories only, so code in a repository listed here with `active:
        false` is indexed and reachable and still will not appear until a
        request names that repository. This endpoint is the only way to tell
        that case from code the index does not hold.
      responses:
        "200":
          description: All known repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repo"
    post:
      tags: [repos]
      summary: Add a repository (idempotent)
      description: >
        The repository id is derived from name+path, so re-posting the same
        repository is idempotent: only its definition (name, source, url, path,
        branch) is updated. Lifecycle state — status, indexed_at, last_commit
        and any in-progress indexing claim — is preserved.

        Requires the admin scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRepoRequest"
      responses:
        "201":
          description: Repository created or re-registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}:
    get:
      tags: [repos]
      summary: Get a repository
      parameters:
        - $ref: "#/components/parameters/RepoID"
      responses:
        "200":
          description: Repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repo"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [repos]
      summary: Delete a repository and all its indexed data
      description: Requires the admin scope.
      parameters:
        - $ref: "#/components/parameters/RepoID"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/repos/{id}/index:
    post:
      tags: [repos]
      summary: Start background indexing of a repository
      description: >
        In single-instance mode indexing has already started when this returns
        (`status: indexing`). With indexes.distributed enabled the request is
        only queued (`status: queued`) and the response carries the job id and
        its queue state; any instance's worker may execute it. At most one
        pending job per repository exists, and a queued job absorbs later
        requests — its force flag is the OR of all of them.

        Requires the admin scope.
      parameters:
        - $ref: "#/components/parameters/RepoID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexRequest"
      responses:
        "202":
          description: Indexing started or job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexAck"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RepoBusy"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}/reset:
    post:
      tags: [repos]
      summary: Clear a stuck indexing claim
      description: >
        Returns a repository stuck in the `indexing` status to `idle`, releasing
        its claim. Claims carry an owner and an expiry and are also released at
        startup; this is the manual escape hatch for a claim whose holder died
        between recoveries.

        Requires the admin scope.
      parameters:
        - $ref: "#/components/parameters/RepoID"
      responses:
        "200":
          description: Repository after the reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repo"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}/jobs:
    get:
      tags: [repos]
      summary: List the repository's queue entries
      description: >
        POST /index and, with indexes.distributed enabled, POST /commits both
        answer 202 whether the work started or was only queued. This endpoint
        is where a client tells those apart: each entry reports its kind
        (`index` or `commits`), its status (`pending`, `running`, `done`,
        `error`), the worker holding it, when it was created, claimed and last
        heartbeated, and the error a failed job recorded. Newest first. Commit
        payloads are never returned.
      parameters:
        - $ref: "#/components/parameters/RepoID"
        - name: limit
          in: query
          required: false
          description: Page size, newest first (default 50, capped at 200)
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Queue entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}/jobs/{job_id}:
    get:
      tags: [repos]
      summary: Get one queue entry
      description: >
        Follows the `job_id` returned by POST /index in distributed mode. A job
        belonging to a different repository is reported as `not_found`.
      parameters:
        - $ref: "#/components/parameters/RepoID"
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexJob"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}/commits:
    post:
      tags: [repos]
      summary: Push commits for incremental indexing
      description: |
        Ingests a batch of commits (with per-file diff statuses) from an
        external client — the source of truth about the repository history.
        Only the affected paths are re-indexed. When the repo already has a
        commit cursor, the first pushed commit must reference it in `parents`;
        otherwise the request is rejected with 409 and the current cursor, and
        nothing is indexed (the client decides to resend the missing range or
        request a full reindex via /index).

        A batch that continues the one still being applied is answered with
        `repo_busy` (retry), not `commit_gap`: the cursor only advances once
        the running batch lands, and `/sync-state.pending_commit` shows what it
        is applying.

        With indexes.distributed enabled the accepted batch is queued as a
        `commits` job (visible under /jobs) instead of being applied by the
        instance that received the request, so an instance that dies after
        answering 202 does not take the batch with it: another worker claims
        it. Batches are applied in the order they were accepted. The gap check
        above still runs before the batch is accepted.

        Ingested paths are subject to the same `repos.ignore` patterns as a
        full index pass, so a push and a full index produce the same result.
        Files the server cannot read are skipped and reported, and the commit
        cursor only advances when the whole batch landed.

        This endpoint carries file contents and therefore has its own, much
        larger body limit (default 64 MiB, `RAGOTA_MAX_COMMIT_BODY_BYTES`)
        instead of the 1 MiB default; an oversized body is a 413 naming the
        limit, not a 400.

        Requires the admin scope.
      parameters:
        - $ref: "#/components/parameters/RepoID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitsRequest"
      responses:
        "202":
          description: Commits accepted; partial indexing started or queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitAck"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            Commit gap (code `commit_gap`) — the first commit does not continue
            the stored cursor — or the repository is busy (code `repo_busy`),
            in which case Retry-After is set.
          headers:
            Retry-After:
              description: Seconds to wait before retrying (repo_busy only)
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}/sync-state:
    get:
      tags: [repos]
      summary: Get the repo's commit cursor, in-flight batch and indexing status
      description: >
        `pending_commit` is the SHA of the batch the system is holding —
        queued or being applied — which is what distinguishes "in flight" from
        "lost". Combine it with `indexed_at` and `last_error` to decide whether
        to resend, and with /jobs to see whether a queued batch has been picked
        up yet.
      parameters:
        - $ref: "#/components/parameters/RepoID"
      responses:
        "200":
          description: Sync state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncState"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/repos/{id}/coverage:
    get:
      tags: [repos]
      summary: Contract coverage of the last full index pass
      description: >
        How much of the repository's outbound contract surface the indexer
        actually resolved, per contract kind: `candidates` counts call sites
        that look like an outbound contract, `edges` counts those that produced
        one.


        It answers the question the graph alone cannot: a CMS that genuinely
        makes 42 HTTP calls and a workflow engine that makes thousands through
        its own helper both look complete when the second one yields 104 edges,
        because a call site that produced no edge leaves nothing behind to
        count. Equal counts mean there was nothing more to find; a large gap
        means the indexer does not understand this project's way of calling
        out, and results built on the graph are incomplete.


        `reported: false` means no coverage-reporting pass has run for this
        repo — the counters are then meaningless rather than zero. Only a full
        index pass writes the summary; a commit-driven partial pass describes
        the files it touched, not the repository. Compare `updated_at` with
        `indexed_at` to spot a summary that predates the current index.
      parameters:
        - $ref: "#/components/parameters/RepoID"
      responses:
        "200":
          description: Coverage summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coverage"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/search:
    post:
      tags: [search]
      summary: Search indexed code (semantic, keyword or hybrid)
      description: >
        Ranked code locations for a question asked in prose — "where does POST
        /cart/checkout go in the frontend", "how is the retry budget
        configured". This and /nav/symbol are the two ways in, and they have
        opposite strengths; choosing the wrong one is the largest avoidable loss
        of answer quality here.


        Measured over 21 benchmark questions against one corpus: this endpoint
        answers a natural-language question at recall@1 0.524, recall@10 0.714,
        MRR 0.587, while /nav/symbol answers a caller that already holds the
        identifier at recall@1 0.667, MRR 0.721 — through its `symbol` field,
        one term matched against a symbol's name or its qualified name. So a
        sentence comes here; an identifier lifted out of a stack trace, a diff
        or a previous answer goes to /nav/symbol. When the answer needs the
        callers of what matched, or the far side of a contract, /context is this
        endpoint with the code graph expanded around every hit.


        `mode` is `hybrid` (the default: both indexes retrieved and fused),
        `keyword` (BM25 alone) or `semantic` (embeddings alone, and only where
        an embedding model is configured). An unrecognized mode is rejected with
        400, never quietly downgraded. The `mode` in the response is the mode
        that ran, so a request that omitted it reads `hybrid` back rather than
        an empty string.


        `repos` decides which repositories answer, and it changes what an empty
        result means. Sent nothing, this endpoint answers from the repositories
        the deployment currently has active, so zero hits means either the index
        does not hold the answer or the repository holding it is not active —
        /repos shows which, and naming that repository in `repos` reads it
        whether it is active or not. Sent a list, zero hits is what those
        repositories have to say and asking again the same way will not change
        it. The code graph is never scoped this way: /services, /topics and the
        /graph endpoints see every indexed repository, so they can name a
        service whose code an unscoped search here will not return.


        `diagnostics` asks the response to carry how the answer was produced:
        which indexes contributed, and whether one of them was unavailable. It
        is off by default and worth sending when a thin or empty result would
        otherwise be taken as the corpus's answer — `diagnostics.degraded` is
        what separates a real zero from a retrieval backend that was down.


        `limit` defaults to 10 and is clamped to at most 100 — an out-of-range
        value is clamped, not rejected. `total` counts what this response
        retrieved rather than corpus-wide matches, so it equals the length of
        `hits` unless `truncated` reports that `max_bytes` dropped some.


        `intent` decides whether the query describes code to find or asks what
        uses it; see the field in SearchRequest. `filter` narrows by language,
        kind and path prefix before ranking. `snippet` and `max_bytes` bound the
        response, and nothing else does: a limit of 20 has measured 34 KB, most
        of it snippet, which is a large bite out of the context window of the
        model that asked.


        Ranking is by retrieval score. Unlike /nav/symbol, this endpoint applies
        no exact-over-substring tiering and does not demote generated or test
        files — a query whose answer is a known identifier will meet the
        generated stubs that share its name, which is the other reason to send
        such a query to /nav/symbol.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search hits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/nav/definition:
    post:
      tags: [nav]
      summary: Find the definition at a file position
      description: >
        The symbol defined at a point in a file: the innermost unit whose line
        range contains `position.line`. Reach for it when a file and a line are
        already in hand — a stack frame, a diff hunk, a hit from /search. When
        an identifier is in hand and the file is not, /nav/symbol is the call:
        it needs no position.


        Lines are 1-based. Line 0 always answers empty, so a caller converting
        from a 0-based editor position must add one. `position.character` is
        accepted and ignored — resolution is by line. Nothing defined at that
        line is an answer rather than an error: 200 with an empty object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefinitionRequest"
      responses:
        "200":
          description: Definition (empty object when not found)
          content:
            application/json:
              schema:
                type: object
                properties:
                  definition:
                    $ref: "#/components/schemas/ASTSymbol"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/nav/references:
    post:
      tags: [nav]
      summary: Find references to the symbol at a file position
      description: >
        Where the symbol at a file position is used. The position resolves the
        same way as /nav/definition — the innermost unit containing
        `position.line`, 1-based — and the answer is the code graph edges that
        point at that unit. Any edge kind can appear, not only direct calls:
        imports, HTTP and RPC calls, produces and consumes.


        Two lookups are concatenated: edges already resolved to this unit, and
        edges in the same repository that name it but were never resolved
        (`target` is then empty). `limit` is the size of the whole answer, not
        of each lookup — it defaults to 50 and is clamped to at most 500. The
        resolved edges are taken first and up to the full budget; the
        unresolved ones only fill what is left, so a hot symbol's near misses
        can never crowd out a reference that actually points at it.


        `word` is what the edge names — a callee name, or a contract key such as
        `topic:orders.created` — not a token read at the cursor. `kind` is the
        edge kind, and `repo_id`/`file_path`/`line` locate the referencing site,
        not the definition.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferencesRequest"
      responses:
        "200":
          description: References
          content:
            application/json:
              schema:
                type: object
                properties:
                  references:
                    type: array
                    items:
                      $ref: "#/components/schemas/ASTReference"
                  total:
                    type: integer
                    description: Length of `references`, not a corpus-wide count
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/graph/neighbors:
    post:
      tags: [graph]
      summary: Edges around a code unit (in/out, with far-side units)
      description: >
        The edges immediately around one unit: `out` is what it calls, produces
        or writes, `in` is what points at it. Each hop carries the unit on the
        far side when the edge resolved; `unit` is absent for an edge naming
        something not indexed here, which is the normal shape of a call into a
        third-party library.


        `unit_id` comes from another answer — a /context item's `unit`, a
        /nav/symbol result, a service's `unit_id`. There is no way to name a
        unit by hand, and ids do not survive a reindex, so they are worth
        following rather than storing. An unknown id is 404.


        Neither list is capped: a unit that everything calls returns every edge
        it has. When the question is about code rather than about one unit's
        wiring, /context expands the graph around search hits instead and
        bounds what it returns.


        Edges are followed across every indexed repository. The active set that
        scopes an unscoped /search does not scope this, so a far-side `unit` may
        sit in a repository that search would not have answered from.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [unit_id]
              properties:
                unit_id:
                  type: string
      responses:
        "200":
          description: Center node with incoming and outgoing edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeighborsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/services/export:
    get:
      tags: [services]
      summary: Service graph as a Mermaid or DOT diagram
      description: >
        The same graph as /services, rendered as diagram text for a human to
        look at. Ask /services instead for a graph to reason over: this answers
        `text/plain`, and the edge labels carry neither `count` nor
        `confidence`.


        One node per repository-and-service pair, labelled with the bare service
        name, or `repo/service` where several repositories share the name. Edges
        are labelled by contract — `kafka <topic>`, `gRPC <method>`,
        `<METHOD> /path`, or `runtime`. There is no grouping by repository in
        either format.


        `format` is `mermaid` (the default) or `dot`; any other value renders
        Mermaid. Unlike /services this endpoint takes no `repo` or `limit`, so
        it always renders the whole estate.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [mermaid, dot]
            default: mermaid
      responses:
        "200":
          description: Diagram text
          content:
            text/plain:
              schema:
                type: string
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Applied to /api/v1 routes when server.auth.type is api_key.

        A key carries a scope, written as a prefix in server.auth.api_keys:
        "read:<key>" may call the retrieval and inspection routes, "admin:<key>"
        may also call every mutating and administrative one (marked "requires
        the admin scope" below). A key with no prefix is admin, which is what
        every key granted before scopes existed — an existing config keeps
        working unchanged, and restricting a key is an edit that opts in. The
        prefix is not part of the credential: a client sends the key itself.

        A key without the scope a route needs gets 403 `forbidden`. When
        server.auth.type is none there are no keys and no scope checks.
    BearerAuth:
      type: http
      scheme: bearer
      description: "Alternative to X-API-Key (Authorization: Bearer <key>)."
...
  responses:
    BadRequest:
      description: Invalid request (code `validation_failed` or `invalid_path`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    GitWebhookPayload:
      type: object
      description: GitHub or GitLab push event (only repository identifiers are used)
      properties:
        repository:
          type: object
          properties:
            name:
              type: string
            full_name:
              type: string
            clone_url:
              type: string
            git_http_url:
              type: string
            ssh_url:
              type: string
        project:
          type: object
          properties:
            name:
              type: string
            path_with_namespace:
              type: string
            git_http_url:
              type: string
            git_ssh_url:
              type: string